Backend/requirements.txt
altair>=6.0.0
attrs>=25.4.0
blinker>=1.9.0
cachetools>=6.2.4
certifi>=2026.1.4
charset-normalizer>=3.4.4
click>=8.3.1
colorama>=0.4.6
# git dependencies removed for stability on HF

greenlet>=3.3.0
idna>=3.11
Jinja2>=3.1.6
jsonschema>=4.26.0
jsonschema-specifications>=2025.9.1
MarkupSafe>=3.0.3
narwhals>=2.15.0
numpy>=1.26.4
packaging>=25.0
pandas>=2.3.3
pillow>=12.1.0
protobuf>=6.33.2
pyarrow>=22.0.0
pydeck>=0.9.1
pyee>=13.0.0
python-dateutil>=2.9.0.post0
pytz>=2025.2
referencing>=0.37.0
requests>=2.32.5
rpds-py>=0.30.0
six>=1.17.0
smmap>=5.0.2
streamlit>=1.53.1
tenacity>=9.1.2
toml>=0.10.2
tornado>=6.5.4
typing_extensions>=4.15.0
tzdata>=2025.3
urllib3>=2.6.2
watchdog>=6.0.0
rich
stable-baselines3>=2.7.1
gymnasium>=1.2.3
tensorboard>=2.20.0
fastapi>=0.135.0
uvicorn>=0.41.0
pydantic>=2.12.5
# keyboard removed as it's not suitable for headless environments

psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
groq>=0.13.0
google-genai
openai
xgboost
scikit-learn
catboost
lightgbm